+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Building access APIs. [#48](https://github.com/rokwire/health-building-block/issues/48)
 
 ## [1.19.0] - 2020-09-24
 ### Added